test: Set annotations in a single kubectl call
This batches kubectl calls so that both fields are annotated at the same
time. This will avoid us seeing log lines where one or the other
annotation may be missing in CI.

I also couldn't help myself cleaning up the test code to use const and
var blocks so it makes the tests easier to read.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>