remove calls to GetAddressSetHashNames() in egressFirewalls
working on removing the dependancy of egressFirewalls on the namespace
struct, so remove the need to grab the lock to get the addressSet name
and just generate the addressSet names.

This is safe because egressFirewall does not add or remove ip's from the
addressSets and attempting to match on addressSet that does not exist
just doesn't match.

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>